requirements.txt
ccxt
requests
pandas
numpy
python-dotenv
loguru
pandas
ta